tools/docs_bridge/requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
grpcio
grpcio-tools
flask
openai-whisper==20231117
tqdm>=4.66